api: Consolidate types.go with operator_types.go
